Fixed all compiler errors with one exception
the typescript compiler does NOT like the lack of an import
statement in kaitaiWorker.ts for KaitaiStream. If I try to include
it, it compiles without issue but it breaks when it attempts to
load in the browser because it tries to use requirejs to actually
import it. This doesn't work.

Everything else however, should work.