Disable font ligatures in hex viewer (#135)
With ligatures enabled, certain combinations of characters, like "ff",
may render with a width less than that of any two other characters, even
when using a monospace font. This violates the assumptions used to align
the hex view and makes lines containing them look ugly, especially when
the ligature is inside the address. To fix this, disable ligatures
entirely in the hex viewer.