Fix temporary error when imports are used, support circular imports
Fixes https://github.com/kaitai-io/kaitai_struct_webide/issues/59

Fixes https://github.com/kaitai-io/kaitai_struct_webide/issues/159 -
this is a particularly notable instance of the bug, since `TypeError:
DosDatetime is not a constructor` is an error that initially everyone
gets when they open the Web IDE in a fresh browser (this is because
`zip.ksy` is selected by default, and it is affected by the bug since it
contains an import).

See https://github.com/kaitai-io/kaitai_struct/issues/1074 and
https://github.com/kaitai-io/kaitai_struct_compiler/pull/264 - the
actual fix was done in the KSC-generated JS code, which was changed to
allow interdependent format modules to be loaded (and added to the
global context in the case of the Web IDE) in any order. However, this
was a backwards-incompatible change, so this set of Web IDE changes just
updates our code to work with the new compiler.