Remove broken unnecessary code for `ioOffset` adjustment
The `if (result.start === childIoOffset) { ... }` code was intentionally
removed, because it turned out to be broken. For example, consider the
following .ksy snippet:

```ksy
meta:
  id: subtype_in_substream
seq:
  - id: foo
    type: u1
  - id: sub
    size: 7
    type: subtype
types:
  subtype:
    seq:
      - id: bar
        type: u1
      - id: body
        type: body_type
  body_type:
    seq:
      - id: qux
        size: 3
```

If you select the `sub` field in the Web IDE's object tree, the
highlighted interval in the hex dump is `[1, 8)` (from 1 inclusive to 8
exclusive). The interval of `sub.bar` is `[1, 2)`, as expected. But the
highlighted interval for `sub.body` is `[1, 4)`, which is wrong - this
would suggest that `sub.bar` and `sub.body` overlap, which is not the
case. The correct interval for `sub.body` should the same as for
`sub.body.qux`, which is `[2, 5)`.

This incorrect highlighting is exactly the result of the `if
(result.start === childIoOffset) { ... }` code, which confuses the
`start` position of `body` of `1` with the start of a new substream,
because `1` also happens to be the `byteOffset` of the current stream.

I believe that removing this piece of code has no observable negative
effect, because a project search shows that `ioOffset` is always used
like `exp.ioOffset + exp.start`, so it doesn't seem to matter whether
`ioOffset` and `start` (or `end`) have "correct" values individually,
because in the end only their sum is used and interpreted.