Display the last primitive array item whose parsing failed
The idea behind the implementation is that if `_debug[*].arr` is
available, it will be more complete, since it will also contain an entry
for the last array item whose value could not even be "constructed" and
inserted into the array (which is the case for primitive types, but one
can also imagine other cases, for example when `size` of the repeated
field is negative or exceeds EOF).

If `_debug[*].arr` doesn't exist (which happens for value instances
containing arrays), we fall back to the previous implementation, except
that we know that no debug info is available.