Accessibility: open file right after selecting it
This is necessary to be able to control the file tree using the
keyboard. Before this commit, the file tree could be navigated using the
arrow keys, but when selecting a file by pressing Enter, nothing
happened other than highlighting the file tree item with a blue
background. It was simply impossible to open a file without a mouse,
because you need a mouse to trigger a double-click event. Now a file
will open as soon as the corresponding file tree entry is selected, so
pressing Enter opens the file.

A side effect is that a single click (instead of a double-click) is now
enough to open a file. I don't think this is a problem - it matches the
behavior of the file tree in Visual Studio Code, for example.