Use `a11y-dialog` for dialogs to fix accessibility issues
The [modal](https://getbootstrap.com/docs/3.4/javascript/#modals)
component in Bootstrap 3.4.1 that we were using to display
modals/dialogs until now had some weird behavior in terms of moving
focus using the keyboard. When moving the focus forward using the Tab
key, it would cycle through the links inside the About modal as
expected. Except that after the Close button, the focus would go to a
"hidden place" (which was actually the entire `div.modal` element)
before the first link in the modal would be focused again. When moving
the focus backward using Shift+Tab, the focus would move from the first
link to the browser address bar, not to the last element (the Close
button in the case of the About modal) as expected.

The [`a11y-dialog`](https://a11y-dialog.netlify.app/) library is
specifically designed to fix accessibility issues like this, so now we
only use some of Bootstrap styles for modal contents, but we no longer
use any Bootstrap JS code to display the modal. Instead, showing/hiding
the modal is now left entirely to `a11y-dialog`.