Make context menus more accessible
`tabindex="-1"` has been removed from context menu items, allowing
navigation between them using the Tab key.

Context menus are now wrapped in a "dialog window" using the
`a11y-dialog` library. This ensures that the Tab key (and the Shift+Tab
key combination) only moves the focus between items in the currently
open context menu. The Esc key closes the context menu. When the context
menu is opened, the entire browser window is covered with a transparent
layer, and clicking on it also closes the context menu.