Fix error when handling a non-object error value
Sometimes our JS runtime library throws a string instead of an `Error`
object - for example, see
https://github.com/kaitai-io/kaitai_struct_javascript_runtime/blob/ca9ba81f19cbd45c2801facaef0c4786cf177bb8/KaitaiStream.ts#L625

This is not ideal and should be fixed. However, it should not cause an
internal error in Web IDE like this:

```
TypeError: Cannot use 'in' operator to search for 'getMessage__T' in End of stream reached, but no terminator 124 found
```

This commit fixes that.