Exclude `toString` from the object tree and JSON dump
Resolves https://github.com/kaitai-io/kaitai_struct_webide/issues/155

This fix implies that if you defined `to_string` as an attribute, it
would be excluded too. However, I don't think there is anything wrong
with that. Although JavaScript lets you define an attribute called
`toString` that is not a function, it's not a good idea because then any
attempt to convert the object to a string will fail. Therefore,
`to_string` should ideally be treated as a reserved keyword, which
brings us to https://github.com/kaitai-io/kaitai_struct/issues/90 again.